finding a platform compatibility issue
